backend/requirements.txt
# ── LegalSarthi Dependencies ──

# Framework
fastapi>=0.115.0,<1.0.0
uvicorn[standard]>=0.34.0,<1.0.0

# MongoDB
motor>=3.6.0,<4.0.0
pymongo>=4.9,<4.10

# Auth & Security
python-jose[cryptography]>=3.3.0,<4.0.0
passlib[bcrypt]>=1.7.4,<2.0.0
bcrypt>=4.0.0,<5.0.0

# Validation
pydantic[email]>=2.10.0,<3.0.0
pydantic-settings>=2.7.0,<3.0.0
email-validator>=2.0.0,<3.0.0

# Gemini AI (new official SDK)
google-genai>=1.0.0

# Utilities
python-dotenv>=1.0.0,<2.0.0
python-multipart>=0.0.17,<1.0.0
